+++ CHANGES.rst
@@ -1,6 +1,12 @@
 Change log
 ==========
 
+**2018-03-13 (1.1.3)**
+
+* Remove some prefixes from address name (issue #13).
+* Put image links in the address.osm file (issue #14).
+* Option to download only the Cadastre files (issue #16).
+
 **2018-03-02 (1.1.2)**
 
 * Remove upload=yes parameter from OSM josm files (issue #12)